+++ modules/features/settings/src/main/java/au/com/shiftyjelly/pocketcasts/settings/viewmodel/AutoDownloadSettingsViewModel.kt
@@ -1,7 +1,6 @@
 package au.com.shiftyjelly.pocketcasts.settings.viewmodel
 
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsEvent
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsTracker
 import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
@@ -10,11 +9,12 @@ import au.com.shiftyjelly.pocketcasts.preferences.Settings
 import au.com.shiftyjelly.pocketcasts.repositories.download.DownloadManager
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.PodcastManager
 import dagger.hilt.android.lifecycle.HiltViewModel
+import io.reactivex.Single
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.schedulers.Schedulers
 import javax.inject.Inject
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 
 @HiltViewModel
@@ -28,14 +28,7 @@ class AutoDownloadSettingsViewModel @Inject constructor(
     override val coroutineContext = Dispatchers.Default
     private var isFragmentChangingConfigurations: Boolean = false
 
-    private var _hasEpisodesWithAutoDownloadEnabled: MutableStateFlow<Boolean> = MutableStateFlow(false)
-    val hasEpisodesWithAutoDownloadEnabled: StateFlow<Boolean> = _hasEpisodesWithAutoDownloadEnabled
-
-    init {
-        viewModelScope.launch(Dispatchers.IO) {
-            _hasEpisodesWithAutoDownloadEnabled.value = podcastManager.hasEpisodesWithAutoDownloadStatus(Podcast.AUTO_DOWNLOAD_NEW_EPISODES)
-        }
-    }
+    suspend fun hasEpisodesWithAutoDownloadEnabled() = podcastManager.hasEpisodesWithAutoDownloadStatus(Podcast.AUTO_DOWNLOAD_NEW_EPISODES)
 
     fun onShown() {
         if (!isFragmentChangingConfigurations) {
@@ -56,8 +49,6 @@ class AutoDownloadSettingsViewModel @Inject constructor(
     }
 
     fun getAutoDownloadUpNext() = settings.autoDownloadUpNext.value
-
-    fun getAutoDownloadNewEpisodes() = settings.autoDownloadNewEpisodes.value
 
     fun getLimitDownload() = settings.autoDownloadLimit.value
 
@@ -125,6 +116,13 @@ class AutoDownloadSettingsViewModel @Inject constructor(
     suspend fun updateAllAutoDownloadStatus(status: Int) {
         podcastManager.updateAllAutoDownloadStatus(status)
     }
+
+    fun countPodcastsAutoDownloading(): Single<Int> = podcastManager.countDownloadStatusRxSingle(Podcast.AUTO_DOWNLOAD_NEW_EPISODES)
+        .subscribeOn(Schedulers.io())
+
+    fun countPodcasts(): Single<Int> = podcastManager.countSubscribedRxSingle()
+        .observeOn(AndroidSchedulers.mainThread())
+        .subscribeOn(Schedulers.io())
 }
 
 fun Boolean.toAutoDownloadStatus(): Int = when (this) {