+++ modules/services/servers/src/main/java/au/com/shiftyjelly/pocketcasts/servers/sync/SyncService.kt
@@ -104,7 +104,10 @@ interface SyncService {
     fun episodeProgressSync(@Header("Authorization") authorization: String, @Body request: EpisodeSyncRequest): Completable
 
     @GET("/subscription/status")
-    fun subscriptionStatus(@Header("Authorization") authorization: String): Single<SubscriptionStatusResponse>
+    fun subscriptionStatusRxSingle(@Header("Authorization") authorization: String): Single<SubscriptionStatusResponse>
+
+    @GET("/subscription/status")
+    suspend fun subscriptionStatus(@Header("Authorization") authorization: String): SubscriptionStatusResponse
 
     @POST("/subscription/purchase/android")
     fun subscriptionPurchase(@Header("Authorization") authorization: String, @Body request: SubscriptionPurchaseRequest): Single<SubscriptionStatusResponse>