+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 7.80
 -----
-
+*   New Features
+    *   Add support for podcast descriptions with HTML formatting 
+        ([#3249](https://github.com/Automattic/pocket-casts-android/pull/3249))
+*   Bug Fixes
+    *   Fix global auto download setting was incorrectly overriding the podcast auto download setting
+        ([#3342](https://github.com/Automattic/pocket-casts-android/pull/3342))
+    *   Fix end of year translation
+        ([#3387](https://github.com/Automattic/pocket-casts-android/pull/3387))
+    *   Fix sleep timer was not stopping as expected
+        ([#3377](https://github.com/Automattic/pocket-casts-android/pull/3377))
+*   Updates
+    *   Add the ability to dismiss the low storage banner in download screen
+        ([#3385](https://github.com/Automattic/pocket-casts-android/pull/3385))
 
 7.79
 -----